Align with the java8 3.0.0 version
