+++ src/main/java/org/apache/datasketches/memory/Positional.java
@@ -40,21 +40,12 @@ public interface Positional {
 
   /**
    * Increments the current position by the given increment.
-   * Asserts that the resource is valid and that the positional invariants are not violated,
-   * otherwise, if asserts are enabled throws an {@link AssertionError}.
+   * Checks that the resource is valid and that the positional invariants are not violated.
    * @param increment the given increment
    * @return BaseBuffer
+   * @throws BufferPositionInvariantsException if positional invariants have been violated.
    */
   Positional incrementPosition(long increment);
-
-  /**
-   * Increments the current position by the given increment.
-   * Checks that the resource is valid and that the positional invariants are not violated,
-   * otherwise throws an {@link IllegalArgumentException}.
-   * @param increment the given increment
-   * @return BaseBuffer
-   */
-  Positional incrementAndCheckPosition(final long increment);
 
   /**
    * Gets the end position
@@ -95,8 +86,7 @@ public interface Positional {
 
   /**
    * Sets the current position.
-   * Asserts that the positional invariants are not violated,
-   * otherwise, if asserts are enabled throws an {@link AssertionError}.
+   * Checks that the positional invariants are not violated.
    * @param position the given current position.
    * @return BaseBuffer
    * @throws BufferPositionInvariantsException if positional invariants have been violated.
@@ -104,18 +94,8 @@ public interface Positional {
   Positional setPosition(long position);
 
   /**
-   * Sets the current position.
-   * Checks that the positional invariants are not violated,
-   * otherwise, throws an {@link IllegalArgumentException}.
-   * @param position the given current position.
-   * @return BaseBuffer
-   */
-  Positional setAndCheckPosition(long position);
-
-  /**
    * Sets start position, current position, and end position.
-   * Asserts that the positional invariants are not violated,
-   * otherwise, if asserts are enabled throws an {@link AssertionError}.
+   * Checks that the positional invariants are not violated.
    * @param start the start position in the buffer
    * @param position the current position between the start and end
    * @param end the end position in the buffer
@@ -123,16 +103,5 @@ public interface Positional {
    * @throws BufferPositionInvariantsException if positional invariants have been violated.
    */
   Positional setStartPositionEnd(long start, long position, long end);
-
-  /**
-   * Sets start position, current position, and end position.
-   * Checks that the positional invariants are not violated,
-   * otherwise, throws an {@link IllegalArgumentException}.
-   * @param start the start position in the buffer
-   * @param position the current position between the start and end
-   * @param end the end position in the buffer
-   * @return BaseBuffer
-   */
-  Positional setAndCheckStartPositionEnd(long start, long position, long end);
 
 }
