+++ src/main/java/org/apache/datasketches/memory/internal/PositionalImpl.java
@@ -19,6 +19,7 @@
 
 package org.apache.datasketches.memory.internal;
 
+import org.apache.datasketches.memory.BufferPositionInvariantsException;
 import org.apache.datasketches.memory.MemoryRequestServer;
 import org.apache.datasketches.memory.Positional;
 
@@ -59,13 +60,9 @@ abstract class PositionalImpl extends ResourceImpl implements Positional {
 
   @Override
   public final PositionalImpl incrementPosition(final long increment) {
-    incrementAndAssertPositionForRead(pos, increment);
-    return this;
-  }
-
-  @Override
-  public final PositionalImpl incrementAndCheckPosition(final long increment) {
-    incrementAndCheckPositionForRead(pos, increment);
+    final long newPos = getPosition() + increment;
+    checkInvariants(start, newPos, end, capacity);
+    pos = newPos;
     return this;
   }
 
@@ -102,13 +99,6 @@ abstract class PositionalImpl extends ResourceImpl implements Positional {
 
   @Override
   public final PositionalImpl setPosition(final long position) {
-    assertInvariants(start, position, end, capacity);
-    pos = position;
-    return this;
-  }
-
-  @Override
-  public final PositionalImpl setAndCheckPosition(final long position) {
     checkInvariants(start, position, end, capacity);
     pos = position;
     return this;
@@ -116,16 +106,6 @@ abstract class PositionalImpl extends ResourceImpl implements Positional {
 
   @Override
   public final PositionalImpl setStartPositionEnd(final long start, final long position,
-      final long end) {
-    assertInvariants(start, position, end, capacity);
-    this.start = start;
-    this.end = end;
-    pos = position;
-    return this;
-  }
-
-  @Override
-  public final PositionalImpl setAndCheckStartPositionEnd(final long start, final long position,
       final long end) {
     checkInvariants(start, position, end, capacity);
     this.start = start;
@@ -135,42 +115,6 @@ abstract class PositionalImpl extends ResourceImpl implements Positional {
   }
 
   //RESTRICTED
-  //Position checks are only used for Buffers
-  //asserts are used for primitives, not used at runtime
-  final void incrementAndAssertPositionForRead(final long position, final long increment) {
-    final long newPos = position + increment;
-    assertInvariants(start, newPos, end, capacity);
-    pos = newPos;
-  }
-
-  //checks are used for arrays and apply at runtime
-  final void incrementAndCheckPositionForRead(final long position, final long increment) {
-    final long newPos = position + increment;
-    checkInvariants(start, newPos, end, capacity);
-    pos = newPos;
-  }
-
-  /**
-   * The invariants equation is: {@code 0 <= start <= position <= end <= capacity}.
-   * If this equation is violated and assertions are enabled,
-   * an <i>AssertionError</i> will be thrown.
-   * @param start the lowest start position
-   * @param pos the current position
-   * @param end the highest position
-   * @param cap the capacity of the backing buffer.
-   */
-  static final void assertInvariants(final long start, final long pos, final long end,
-      final long cap) {
-    assert (start | pos | end | cap | (pos - start) | (end - pos) | (cap - end) ) >= 0L
-        : "Violation of Invariants: "
-        + "start: " + start
-        + " <= pos: " + pos
-        + " <= end: " + end
-        + " <= cap: " + cap
-        + "; (pos - start): " + (pos - start)
-        + ", (end - pos): " + (end - pos)
-        + ", (cap - end): " + (cap - end);
-  }
 
   /**
    * The invariants equation is: {@code 0 <= start <= position <= end <= capacity}.
@@ -183,7 +127,7 @@ abstract class PositionalImpl extends ResourceImpl implements Positional {
   static final void checkInvariants(final long start, final long pos, final long end,
         final long cap) {
     if ((start | pos | end | cap | (pos - start) | (end - pos) | (cap - end) ) < 0L) {
-      throw new IllegalArgumentException(
+      throw new BufferPositionInvariantsException(
           "Violation of Invariants: "
               + "start: " + start
               + " <= pos: " + pos
